==> code/3.1.diff_expression_visualization.Rmd <==
---
title: "MitoSignatures"
author: "Isabel Duarte"
date: "2023-06-10"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=TRUE, message=FALSE, warnings=FALSE, cache = TRUE)

library(here)
library(tidyverse)
library(mitocarta)
library(patchwork)

```



# NOTE: Cannot knit this script in < 200 GB RAM machines 

This script has been superseded by its version 2.
There is no need to run this script!




## Function to plot the most extreme DEGs

```{r boxplot_funct, eval=FALSE}

## Function to plot the counts for extreme LFC differentially expressed genes

plot_extreme_degs <- function (diff_exprs_list,
                               n_degs_to_plot = 20,
                               mito_yes_no = "no") {
  
  #### PLOTING FUNCTIONS | START ####
  boxplot_organ <-
    function (counts_data, my_title = "Boxplot per Gene") {
      ggplot(counts_data, aes(x = organ, y = log2(counts + 1))) +
        geom_boxplot(aes(color = metabolic_rate)) +
        labs(x = "Organs", y = "Log2 Gene Counts", title = my_title) +
        theme(axis.text.x = element_text(
          angle = 90,
          vjust = 0.5,
          hjust = 1
        )) +
        theme(legend.position = "bottom") +
        facet_wrap(vars(gene_id), scales = "free_y")
    }
  
  boxplot_organ_and_cancer_status <-
    function (counts_data, my_title = "Boxplot per Gene and Cancer Status") {
      ggplot(counts_data, aes(
        x = interaction(organ, cancer_status),
        y = log2(counts + 1)
      )) +
        geom_boxplot(aes(color = metabolic_rate)) +
        labs(x = "Organs", y = "Log2 Gene Counts", title = my_title) +
        theme(axis.text.x = element_text(
          angle = 90,
          vjust = 0.5,
          hjust = 1
        )) +
        theme(legend.position = "bottom") +
        facet_wrap(vars(gene_id), scales = "free_y")
    }
  #### PLOTING FUNCTIONS | END ####
  
  
  #### Main Function ####
  
  # Create list to hold objets from this analysis
  deg_analysis <- list()
  
  # Get the dataset name
  dataset_name = rlang::as_label(rlang::enexpr(diff_exprs_list))
  
  ###
  ### Filter mitochondrial genes only
  ###
  
  if (mito_yes_no == "yes") {
    ### Mitocarta Gene IDs
    # Get all mitocarta ensembl ids (including the ones with several ids separated by pipe)
    deg_analysis$mitocarta_gene_ids <-
      unlist(
        strsplit(
          mitocarta::mitocarta_data$A_Human_MitoCarta3$EnsemblGeneID_mapping_version_20200130,
          "\\|"
        )
      )
    
    ### Mitochondrial DEGs in toptable data
    # Get all toptable ensembl ids without the .version (because mitocarta does not have it)
    deg_analysis$toptable_gene_ids <-
      gsub(
        pattern = "(ENSG\\d+)(\\.\\d+)?",
        replacement = "\\1",
        row.names(diff_exprs_list$topGenes$table)
      )
    
    # Filter toptable_gene_ids to include only mitocarta genes
    deg_analysis$mito_toptable_gene_ids <-
      deg_analysis$toptable_gene_ids[deg_analysis$toptable_gene_ids %in% deg_analysis$mitocarta_gene_ids]
    
    # Slice DEGS toptable to include only Mitochondrial top/bottom n_degs_to_plot, sorted by LFC
      # Important note: Subseting matrices using [] does partial matching for the rownames 
      # (so it does not match the .version)
    
    # Top genes with negative Log Fold Change
    deg_analysis$low_exprs <-
      diff_exprs_list$topGenes$table[deg_analysis$mito_toptable_gene_ids, ] %>%
      dplyr::arrange(logFC) %>%
      dplyr::slice_head (n = n_degs_to_plot)
    
    # Top genes with positive Log Fold Change
    deg_analysis$high_exprs <-
      diff_exprs_list$topGenes$table[deg_analysis$mito_toptable_gene_ids, ] %>%
      dplyr::arrange(logFC) %>%
      dplyr::slice_tail (n = n_degs_to_plot)
  }
  
  ##
  ## Do not filter mitochondrial genes
  ##
  
  if (mito_yes_no == "no") {
    ### Slice DEGS toptable to include only the top/bottom n_degs_to_plot, sorted by LFC
    # Top genes with negative Log Fold Change
    deg_analysis$low_exprs <- diff_exprs_list$topGenes$table %>%
      dplyr::arrange(logFC) %>%
      dplyr::slice_head (n = n_degs_to_plot)
    
    # Top genes with positive Log Fold Change
    deg_analysis$high_exprs <- diff_exprs_list$topGenes$table %>%
      dplyr::arrange(logFC) %>%
      dplyr::slice_tail (n = n_degs_to_plot)
  }
  
  ###
  ### Common steps
  ###
  
  # Bind the top high and low DEGs
  deg_analysis$extreme_de <-
    dplyr::bind_rows (deg_analysis$low_exprs, deg_analysis$high_exprs)
  
  
  ### Create  a tidy dataframe with count values and metadata for selected DEGs
  # Add the metadata to the counts table (filtered by the genes with extreme LFC)
  deg_analysis$extreme_de_counts_metadata <-
    diff_exprs_list$data_clean$combined_counts %>%
    filter(gene_id %in% row.names(deg_analysis$extreme_de)) %>%
    pivot_longer(names_to = "sample_id",
                 values_to = "counts",
                 cols = -1) %>%
    left_join(., diff_exprs_list$data_clean$metadata, by = "sample_id")
  
  ###
  ### Plot counts for selected DEGs
  ###
  
  # Boxplot per organ faceted by gene
  boxplot_organ(
    counts_data = deg_analysis$extreme_de_counts_metadata,
    my_title = ifelse(
      mito_yes_no == "yes",
      "Extreme Mitochondrial DEGs",
      "Extreme DEGs"
    )
  ) ->
    deg_analysis$plots$p1_box_gene
  
  # Boxplot per organ per cancer status faceted by gene
  boxplot_organ_and_cancer_status(
    counts_data = deg_analysis$extreme_de_counts_metadata,
    my_title = ifelse(
      mito_yes_no == "yes",
      paste0("Extreme Mitochondrial DEGs | ", dataset_name),
      paste0("Extreme DEGs | ", dataset_name)
    )
  ) ->
    deg_analysis$plots$p2_box_gene_cancer
  
  ### Return the list with the analysis and plots
  return(deg_analysis)
}
```

## Boxplots of the most extreme DEGs | All genes & Mitochondrial genes 

```{r boxplots, dependson="degs_plot_function", eval=FALSE}
#
##
### Boxplots
##
#


# Load the DE data
gtex_de <- readRDS(file = here("data/analysis/gtex_de.RDS"))
tcga_de <- readRDS(file = here("data/analysis/tcga_de.RDS"))
combined_de_no_interaction <- readRDS(file = here("data/analysis/combined_de_no_interaction.RDS"))
combined_de_only_interaction <- readRDS(file = here("data/analysis/combined_de_only_interaction.RDS"))
combined_de_with_interaction <- readRDS(file = here("data/analysis/combined_de_with_interaction.RDS"))

# Create the lists to save the data
gtex_de_plots <- list() 
tcga_de_plots <- list() 
combined_de_no_interaction_plots <- list() 
combined_de_only_interaction_plots <- list()
combined_de_with_interaction_plots <- list() 


# Call the function for DEG files for each dataset
gtex_de_plots$all <- 
  plot_extreme_degs(diff_exprs_list = gtex_de ,
                    n_degs_to_plot = 10,
                    mito_yes_no = "no")
tcga_de_plots$all <- 
  plot_extreme_degs(diff_exprs_list = tcga_de,
                    n_degs_to_plot = 10,
                    mito_yes_no = "no")
combined_de_no_interaction_plots$all <- 
  plot_extreme_degs(diff_exprs_list = combined_de_no_interaction,
                    n_degs_to_plot = 10,
                    mito_yes_no = "no")
combined_de_only_interaction_plots <- 
  plot_extreme_degs(diff_exprs_list = combined_de_only_interaction,
                    n_degs_to_plot = 10,
                    mito_yes_no = "no")
combined_de_with_interaction_plots$all <- 
  plot_extreme_degs(diff_exprs_list = combined_de_with_interaction,
                    n_degs_to_plot = 10,
                    mito_yes_no = "no")

# Call the function for DEG files for each dataset only for mitochondrial genes
gtex_de_plots$mito <- 
  plot_extreme_degs(diff_exprs_list = gtex_de ,
                    n_degs_to_plot = 10,
                    mito_yes_no = "yes")
tcga_de_plots$mito <- 
  plot_extreme_degs(diff_exprs_list = tcga_de,
                    n_degs_to_plot = 10,
                    mito_yes_no = "yes")
combined_de_no_interaction_plots$mito <- 
  plot_extreme_degs(diff_exprs_list = combined_de_no_interaction,
                    n_degs_to_plot = 10,
                    mito_yes_no = "yes")
combined_de_only_interaction_plots$mito <- 
  plot_extreme_degs(diff_exprs_list = combined_de_only_interaction,
                    n_degs_to_plot = 10,
                    mito_yes_no = "no")
combined_de_with_interaction_plots$mito <- 
  plot_extreme_degs(diff_exprs_list = combined_de_with_interaction,
                    n_degs_to_plot = 10,
                    mito_yes_no = "yes")

```

## View the plots

```{r view_boxplots, dependson="boxplots", eval=FALSE}

# Boxplots
gtex_de_plots$all$plots$p2_box_gene_cancer
tcga_de_plots$all$plots$p2_box_gene_cancer
combined_de_no_interaction_plots$all$plots$p2_box_gene_cancer
combined_de_only_interaction_plots$all$plots$p2_box_gene_cancer
combined_de_with_interaction_plots$all$plots$p2_box_gene_cancer

gtex_de_plots$mito$plots$p2_box_gene_cancer
tcga_de_plots$mito$plots$p2_box_gene_cancer
combined_de_no_interaction_plots$mito$plots$p2_box_gene_cancer
combined_de_only_interaction_plots$mito$plots$p2_box_gene_cancer
combined_de_with_interaction_plots$mito$plots$p2_box_gene_cancer

```

## Visualize Coefficients of the different datasets and linear models

```{r coef, dependson="boxplots", eval=FALSE}
#
##
### Coefficients | Merge data to compare
##
#

# Create list
lrt_coef_df <- list()

# Get coefficients in data frame format and change the column names to include the dataset name (except the gene_id column)
lrt_coef_df$gtex <- data.frame(gene_id = rownames(gtex_de$lrt$coefficients), gtex_de$lrt$coefficients, row.names = NULL)
colnames (lrt_coef_df$gtex)[-1] <- paste0(colnames(lrt_coef_df$gtex)[-1], "_gtex")

lrt_coef_df$tcga <- data.frame(gene_id = rownames(tcga_de$lrt$coefficients), tcga_de$lrt$coefficients, row.names = NULL)
colnames (lrt_coef_df$tcga)[-1] <- paste0(colnames(lrt_coef_df$tcga)[-1], "_tcga")

lrt_coef_df$combined_no_interaction <- data.frame(gene_id = rownames(combined_de_no_interaction$lrt$coefficients), combined_de_no_interaction$lrt$coefficients, row.names = NULL)
colnames (lrt_coef_df$combined_no_interaction)[-1] <- paste0(colnames(lrt_coef_df$combined_no_interaction)[-1], "_comb_NO")

lrt_coef_df$combined_only_interaction <- data.frame(gene_id = rownames(combined_de_only_interaction$lrt$coefficients), combined_de_only_interaction$lrt$coefficients, row.names = NULL)
colnames (lrt_coef_df$combined_only_interaction)[-1] <- paste0(colnames(lrt_coef_df$combined_only_interaction)[-1], "_comb_ONLY")

lrt_coef_df$combined_with_interaction <- data.frame(gene_id = rownames(combined_de_with_interaction$lrt$coefficients), combined_de_with_interaction$lrt$coefficients, row.names = NULL)
colnames (lrt_coef_df$combined_with_interaction)[-1] <- paste0(colnames(lrt_coef_df$combined_with_interaction)[-1], "_comb_WITH")


# Merge all dataframes by common gene ids
lrt_coef_df$merged <- Reduce(function(x,y) inner_join (x,y, by="gene_id"), lrt_coef_df)


# Pivot the dataframe to make the table tidy
lrt_coef_df$merged_long <- lrt_coef_df$merged %>%
  pivot_longer(cols = !gene_id,
               cols_vary = "slowest",
               names_to = "coeff_name",
               values_to = "coeff_value") %>%
  mutate(coeff_name = factor(coeff_name,
                                levels = colnames(lrt_coef_df$merged)[-1]))

### Plotting

# Boxplots | View the distributions of all coefficients (using only common genes)
lrt_coef_df$merged_long %>%
  ggplot(aes(x=coeff_name, y=coeff_value, fill=coeff_name)) +
  geom_boxplot(show.legend = FALSE) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

# Histograms
lrt_coef_df$merged_long %>%
  group_by(coeff_name) %>%
  ggplot(aes(x=coeff_value, fill=coeff_name)) +
  geom_histogram(show.legend = FALSE) +
  facet_wrap(vars(coeff_name))

```

## Visualize LFC and CPMs of the different datasets and linear models

```{r top_genes, dependson="boxplots", eval=FALSE}

#
##
### Top genes | Merge data to compare
##
#

# Create list
top_genes_df <- list()

# Get top genes table and change the column names to include the dataset name (except the gene_id column)
top_genes_df$gtex <- 
  data.frame(gene_id = rownames(gtex_de$topGenes$table), gtex_de$topGenes$table, row.names = NULL)
colnames (top_genes_df$gtex)[-1] <- paste0(colnames(top_genes_df$gtex)[-1], "_gtex")

top_genes_df$tcga <- 
  data.frame(gene_id = rownames(tcga_de$topGenes$table), tcga_de$topGenes$table, row.names = NULL)
colnames (top_genes_df$tcga)[-1] <- paste0(colnames(top_genes_df$tcga)[-1], "_tcga")

top_genes_df$combined_no_interaction <- 
  data.frame(gene_id = rownames(combined_de_no_interaction$topGenes$table), combined_de_no_interaction$topGenes$table, row.names = NULL)
colnames (top_genes_df$combined_no_interaction)[-1] <- paste0(colnames(top_genes_df$combined_no_interaction)[-1], "_comb_NO")

top_genes_df$combined_only_interaction <- 
  data.frame(gene_id = rownames(combined_de_only_interaction$topGenes$table), combined_de_only_interaction$topGenes$table, row.names = NULL)
colnames (top_genes_df$combined_only_interaction)[-1] <- paste0(colnames(top_genes_df$combined_only_interaction)[-1], "_comb_ONLY")

top_genes_df$combined_with_interaction <- 
  data.frame(gene_id = rownames(combined_de_with_interaction$topGenes$table), combined_de_with_interaction$topGenes$table, row.names = NULL)
colnames (top_genes_df$combined_with_interaction)[-1] <- paste0(colnames(top_genes_df$combined_with_interaction)[-1], "_comb_WITH")


# Merge all dataframes by common gene ids
top_genes_df$merged <- Reduce(function(x,y) inner_join (x,y, by="gene_id"), top_genes_df)


# Pivot LogFC and LogCPM data to make it tidy

top_genes_df$logFC_long <-
  top_genes_df$merged %>%
  select(gene_id, contains("LogFC")) %>%
  pivot_longer(cols = !gene_id, 
               names_to = "dataset_model", 
               values_to = "LogFC",
               cols_vary = "fastest") %>%
  mutate(dataset_model = factor(dataset_model,
                                levels = c("logFC_gtex","logFC_tcga",
                                           "logFC_comb_NO",
                                           "logFC_comb_ONLY",
                                           "logFC_comb_WITH") ) )
  
top_genes_df$logCPM_long <-
  top_genes_df$merged %>%
  select(gene_id, contains("LogCPM")) %>%
  pivot_longer(cols = !gene_id, 
               names_to = "dataset_model", 
               values_to = "LogCPM",
               cols_vary = "fastest") %>%
  mutate(dataset_model = factor(dataset_model,
                                levels = c("logCPM_gtex","logCPM_tcga",
                                           "logCPM_comb_NO",
                                           "logCPM_comb_ONLY",
                                           "logCPM_comb_WITH") ) )

top_genes_df$FDR_long <-
  top_genes_df$merged %>%
  select(gene_id, contains("FDR")) %>%
  pivot_longer(cols = !gene_id, 
               names_to = "dataset_model", 
               values_to = "FDR",
               cols_vary = "fastest") %>%
  mutate(dataset_model = factor(dataset_model,
                                levels = c("FDR_gtex","FDR_tcga",
                                           "FDR_comb_NO",
                                           "FDR_comb_ONLY",
                                           "FDR_comb_WITH") ) )


#### Plotting Log Fold Change

# Boxplots | View the distributions of Log Fold Change values (using only common genes)
top_genes_df$logFC_long %>%
  ggplot(aes(x=dataset_model, y=LogFC, fill=dataset_model)) +
  geom_boxplot(show.legend = FALSE) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

# Histograms
top_genes_df$logFC_long %>%
  group_by(dataset_model) %>%
  ggplot(aes(x=LogFC, fill=dataset_model)) +
  geom_histogram(show.legend = FALSE) +
  facet_wrap(vars(dataset_model))



#### Plotting Log Counts per Million

# Boxplots | View the distributions of Log CPM values (using only common genes)
top_genes_df$logCPM_long %>%
  ggplot(aes(x=dataset_model, y=LogCPM, fill=dataset_model)) +
  geom_boxplot(show.legend = FALSE) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

# Histograms
top_genes_df$logCPM_long %>%
  group_by(dataset_model) %>%
  ggplot(aes(x=LogCPM, fill=dataset_model)) +
  geom_histogram(show.legend = FALSE) +
  facet_wrap(vars(dataset_model))



#######################################
#### HERE | Code under development #### 
#######################################



```
