==> code/3.1.diff_expression_visualization_v2.Rmd <==
---
title: "MitoSignatures"
author: "Isabel Duarte"
date: "2023-06-23"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=TRUE, message=FALSE, warnings=FALSE, cache = FALSE)

library(here)
library(tidyverse)
library(mitocarta)
library(biomaRt)
library(ggpattern)
```

## NOTE | Do NOT KNIT this file because the loading of the RDS files is too slow and the knitting fails

## NOTE 2 | Cannot save the de_boxplots object to disk because it is HUGE (lobstr::obj_size(de_boxplots) is 33.13 GB)

---

## Function to plot the most extreme DEGs

```{r boxplot_funct, eval=TRUE}

#
##
### Function to format data | ALL DEGs
##
#
format_data <- function (diff_exprs_list, n_degs_to_plot=12, mart = ensembl) {
  
  # Create list to hold objects from this analysis
  deg_analysis <- list()
  
  # Get the dataset name
  deg_analysis$dataset_name = rlang::as_label(rlang::enexpr(diff_exprs_list))
  
  ### Slice DEGS toptable to include only the top/bottom n_degs_to_plot, sorted by LFC
  # Top genes with negative Log Fold Change
  deg_analysis$low_exprs <- diff_exprs_list$topGenes$table %>%
    dplyr::arrange(logFC) %>%
    dplyr::slice_head (n = n_degs_to_plot)
  
  # Top genes with positive Log Fold Change
  deg_analysis$high_exprs <- diff_exprs_list$topGenes$table %>%
    dplyr::arrange(logFC) %>%
    dplyr::slice_tail (n = n_degs_to_plot)
  
  
  ### Create  a tidy dataframe with count values and metadata for selected DEGs
  # Add the metadata to the counts table (filtered by the genes with extreme LFC)
  
  ### down regulated genes
  deg_analysis$down_de_counts_metadata <-
    diff_exprs_list$data_clean$combined_counts %>%
    filter(gene_id %in% row.names(deg_analysis$low_exprs)) %>%
    pivot_longer(names_to = "sample_id",
                 values_to = "counts",
                 cols = -1) %>%
    mutate(fc_dir = "down") %>%
    left_join(., diff_exprs_list$data_clean$metadata, by = "sample_id")
  
  ### up regulated genes
  deg_analysis$up_de_counts_metadata <-
    diff_exprs_list$data_clean$combined_counts %>%
    filter(gene_id %in% row.names(deg_analysis$high_exprs)) %>%
    pivot_longer(names_to = "sample_id",
                 values_to = "counts",
                 cols = -1) %>%
    mutate(fc_dir = "up") %>%
    left_join(., diff_exprs_list$data_clean$metadata, by = "sample_id")
  
  # Bind the top up and down regulated genes
  deg_analysis$extreme_de_counts_metadata0 <- 
    dplyr::bind_rows (deg_analysis$up_de_counts_metadata, 
                      deg_analysis$down_de_counts_metadata) %>%
    mutate(gene_id = gsub("\\..*$", "", gene_id))
  
  ### Gene symbols
  # Fetch gene symbols and gene names
  gene_info <- getBM(
    attributes = c("ensembl_gene_id", "external_gene_name"),
    filters = "ensembl_gene_id",
    values = deg_analysis$extreme_de_counts_metadata0$gene_id,
    mart = mart
  )
  
  deg_analysis$extreme_de_counts_metadata <-
    deg_analysis$extreme_de_counts_metadata0 %>%
    left_join(., gene_info, join_by("gene_id" == "ensembl_gene_id"), 
                                    multiple = "first", 
                                    relationship = "many-to-many") %>%
    mutate(
      external_gene_name = coalesce(external_gene_name, gene_id),
      external_gene_name = case_when(external_gene_name == "" ~ gene_id,
                                     .default = external_gene_name) ) %>%
    rename(gene_symbol = external_gene_name)
  
  ### Summary table with Up and Down regulated genes
  deg_analysis$sum_table <- dplyr::bind_rows (rownames_to_column(deg_analysis$low_exprs, var = "gene_id"),
                                              arrange(rownames_to_column(deg_analysis$high_exprs, var = "gene_id"),
                                                      desc(logFC))) %>%
    mutate(gene_id = gsub("\\..*$", "", gene_id)) %>%
    left_join(., gene_info, join_by("gene_id" == "ensembl_gene_id"), 
                                    multiple = "first", 
                                    relationship = "many-to-many") %>%
    mutate(
      external_gene_name = coalesce(external_gene_name, gene_id),
      external_gene_name = case_when(external_gene_name == "" ~ gene_id,
                                     .default = external_gene_name) ) %>%
    rename(gene_symbol = external_gene_name) %>%
    relocate(gene_symbol, .after = gene_id)
  
  ### Return the list with the data formatted for plots
  # return(deg_analysis)
  return(list(extreme_data = deg_analysis$extreme_de_counts_metadata, 
              sum_table = deg_analysis$sum_table))
  
}


#
##
### Function to format data | MITOCHONDRIAL DEGs
##
#
format_data_mito <- function (diff_exprs_list, n_degs_to_plot=12, mart = ensembl, mitocarta_genes = mitocarta_gene_ids) {
  
  # Create list to hold objects from this analysis
  deg_analysis <- list()
  
  # Get the dataset name
  deg_analysis$dataset_name = rlang::as_label(rlang::enexpr(diff_exprs_list))
  
  ### Mitochondrial DEGs in toptable data
  # Get all toptable ensembl ids without the .version (because mitocarta does not have it)
  deg_analysis$toptable_gene_ids <-
    gsub(
      pattern = "(ENSG\\d+)(\\.\\d+)?",
      replacement = "\\1",
      row.names(diff_exprs_list$topGenes$table)
    )
  
  # Filter toptable_gene_ids to include only mitocarta genes
  deg_analysis$mito_toptable_gene_ids <-
    deg_analysis$toptable_gene_ids[deg_analysis$toptable_gene_ids %in% mitocarta_gene_ids]
  
  # Slice DEGS toptable to include only Mitochondrial top/bottom n_degs_to_plot, sorted by LFC
  # Important note: Subsetting matrices using [] does partial matching for the rownames 
  # (so it does not match the .version)
  
  # Top genes with negative Log Fold Change
  deg_analysis$low_exprs <-
    diff_exprs_list$topGenes$table[deg_analysis$mito_toptable_gene_ids, ] %>%
    dplyr::arrange(logFC) %>%
    dplyr::slice_head (n = n_degs_to_plot)
  
  # Top genes with positive Log Fold Change
  deg_analysis$high_exprs <-
    diff_exprs_list$topGenes$table[deg_analysis$mito_toptable_gene_ids, ] %>%
    dplyr::arrange(logFC) %>%
    dplyr::slice_tail (n = n_degs_to_plot)
  
  ### Create  a tidy dataframe with count values and metadata for selected DEGs
  # Add the metadata to the counts table (filtered by the genes with extreme LFC)
  
  ### down regulated genes
  deg_analysis$down_de_counts_metadata <-
    diff_exprs_list$data_clean$combined_counts %>%
    filter(gene_id %in% row.names(deg_analysis$low_exprs)) %>%
    pivot_longer(names_to = "sample_id",
                 values_to = "counts",
                 cols = -1) %>%
    mutate(fc_dir = "down") %>%
    left_join(., diff_exprs_list$data_clean$metadata, by = "sample_id")
  
  ### up regulated genes
  deg_analysis$up_de_counts_metadata <-
    diff_exprs_list$data_clean$combined_counts %>%
    filter(gene_id %in% row.names(deg_analysis$high_exprs)) %>%
    pivot_longer(names_to = "sample_id",
                 values_to = "counts",
                 cols = -1) %>%
    mutate(fc_dir = "up") %>%
    left_join(., diff_exprs_list$data_clean$metadata, by = "sample_id")
  
  # Bind the top up and down regulated genes
  deg_analysis$extreme_de_counts_metadata0 <- 
    dplyr::bind_rows (deg_analysis$up_de_counts_metadata, 
                      deg_analysis$down_de_counts_metadata) %>%
    mutate(gene_id = gsub("\\..*$", "", gene_id))
  
  ### Gene symbols
  # Fetch gene symbols and gene names
  gene_info <- getBM(
    attributes = c("ensembl_gene_id", "external_gene_name"),
    filters = "ensembl_gene_id",
    values = deg_analysis$extreme_de_counts_metadata0$gene_id,
    mart = mart
  )
  
  deg_analysis$extreme_de_counts_metadata <-
    deg_analysis$extreme_de_counts_metadata0 %>%
    left_join(., gene_info, join_by("gene_id" == "ensembl_gene_id"), 
                                    multiple = "first", 
                                    relationship = "many-to-many") %>%
    mutate(
      external_gene_name = coalesce(external_gene_name, gene_id),
      external_gene_name = case_when(external_gene_name == "" ~ gene_id,
                                     .default = external_gene_name) ) %>%
    rename(gene_symbol = external_gene_name)
  
  ### Summary table with Up and Down regulated genes
  deg_analysis$sum_table <- dplyr::bind_rows (rownames_to_column(deg_analysis$low_exprs, var = "gene_id"),
                                              arrange(rownames_to_column(deg_analysis$high_exprs, var = "gene_id"),
                                                      desc(logFC))) %>%
    mutate(gene_id = gsub("\\..*$", "", gene_id)) %>%
    left_join(., gene_info, join_by("gene_id" == "ensembl_gene_id"), 
                                    multiple = "first", 
                                    relationship = "many-to-many") %>%
    mutate(
      external_gene_name = coalesce(external_gene_name, gene_id),
      external_gene_name = case_when(external_gene_name == "" ~ gene_id,
                                     .default = external_gene_name) ) %>%
    rename(gene_symbol = external_gene_name) %>%
    relocate(gene_symbol, .after = gene_id)
  
  ### Return the list with the data formatted for plots
  
  # return(deg_analysis)
  return(list(extreme_data = deg_analysis$extreme_de_counts_metadata, 
              sum_table = deg_analysis$sum_table))
  
}

#
##
### Function for boxplots using the formatted data #### HERE
##
#

plot_metab_and_cancer <-
  function (counts_data, my_title = "") {
    ggplot(counts_data, aes(x = metabolic_rate,
                            y = log2(counts + 1))) +
      geom_boxplot_pattern(aes(fill=metabolic_rate, pattern=cancer_status),
                           pattern_size=0.2, linewidth=0.2, outlier.size = 0.2) +
      labs(x = "Stripes = Healthy | Dots = Cancer", y = "Log2 Gene Counts", title = my_title) +
      theme_light() +
      theme(axis.text.x = element_text(angle = 0, vjust = 0.5, hjust = 0.5, size = 7),
            axis.title.x = element_text(vjust = -3),
            plot.margin = margin(t=1,r=1,b=1,l=1, unit="cm")) +
      theme(legend.position = "none") +
      scale_x_discrete(labels=c(
        "low" = "Low metabolism",
        "high" = "High metabolism")) +
      facet_wrap(vars(gene_symbol), scales = "free_y")
  }

plot_organ_and_cancer <-
  function (counts_data, my_title = "") {
    ggplot(counts_data, aes(x = interaction(organ, cancer_status),
                            y = log2(counts + 1))) +
      # geom_boxplot(aes(fill = metabolic_rate)) +
      # geom_boxplot(aes(color = cancer_status)) +
      geom_boxplot_pattern(aes(fill=metabolic_rate, pattern=cancer_status),
                           pattern_size=0.2, linewidth=0.2, outlier.size = 0.2) +
      stat_summary(mapping = aes(group = cancer_status), fun = mean, geom="line", 
                   linetype="solid", color="gray10", linewidth=0.5) +
      labs(x = "Organs and Cancer status", y = "Log2 Gene Counts", title = my_title) +
      theme_light() +
      theme(axis.text.x = element_text(
        angle = 90, vjust = 0.5, hjust = 1, size = 7)) +
      theme(legend.position = "none") +
      scale_x_discrete(labels=c(
      "bladder.non_cancer" = "Bladder Healthy",
      "colon.non_cancer" = "Colon Healthy",
      "skin.non_cancer" = "Skin Healthy",
        "bladder.cancer" = "Bladder Cancer",
        "colon.cancer" = "Colon Cancer",
        "skin.cancer" = "Skin Cancer",
      "brain.non_cancer" = "Brain Healthy",
      "liver.non_cancer" = "Liver Healthy",
      "kidney.non_cancer" = "Kidney Healthy",
        "brain.cancer" = "Brain Cancer",
        "liver.cancer" = "Liver Cancer",
        "kidney.cancer" = "Kidney Cancer")) +
      facet_wrap(vars(gene_symbol), scales = "free_y")
  }

  plot_organ <-
    function (counts_data, my_title = "Boxplot per Gene") {
      ggplot(counts_data, aes(x = organ, y = log2(counts + 1))) +
        geom_boxplot(aes(color = metabolic_rate)) +
        labs(x = "Organs", y = "Log2 Gene Counts", title = my_title) +
        theme(axis.text.x = element_text(
          angle = 90,
          vjust = 0.5,
          hjust = 1
        )) +
        theme(legend.position = "none") +
        facet_wrap(vars(gene_id), scales = "free_y")
    }

```


```{r load_data, eval=FALSE}

## Load the DE data and keep in memory only the required tables form the full list

#GTEx
gtex_de_full <- readRDS(file = here("data/analysis/gtex_de.RDS"))
gtex_de <- list(data_clean = gtex_de_full$data_clean,
                 topGenes = gtex_de_full$topGenes,
                 lrt = gtex_de_full$lrt)
rm(gtex_de_full)

#TCGA
tcga_de_full <- readRDS(file = here("data/analysis/tcga_de.RDS"))
tcga_de <- list(data_clean = tcga_de_full$data_clean,
                 topGenes = tcga_de_full$topGenes,
                 lrt = tcga_de_full$lrt)
rm(tcga_de_full)

#Combined NO interaction
combined_de_no_interaction_full <- readRDS(file = here("data/analysis/combined_de_no_interaction.RDS"))
combined_de_no_interaction <- list(data_clean = combined_de_no_interaction_full$data_clean,
                 topGenes = combined_de_no_interaction_full$topGenes,
                 lrt = combined_de_no_interaction_full$lrt)
rm(combined_de_no_interaction_full)

#Combined ONLY interaction
combined_de_only_interaction_full <- readRDS(file = here("data/analysis/combined_de_only_interaction.RDS"))
combined_de_only_interaction <- list(data_clean = combined_de_only_interaction_full$data_clean,
                 topGenes = combined_de_only_interaction_full$topGenes,
                 lrt = combined_de_only_interaction_full$lrt)
rm(combined_de_only_interaction_full)

#Combined WITH interaction
combined_de_with_interaction_full <- readRDS(file = here("data/analysis/combined_de_with_interaction.RDS"))
combined_de_with_interaction <- list(data_clean = combined_de_with_interaction_full$data_clean, 
                 topGenes = combined_de_with_interaction_full$topGenes, 
                 lrt = combined_de_with_interaction_full$lrt)
rm(combined_de_with_interaction_full)

```


## Get objects required for functions + Creating required objects and directories

```{r boxplots_prep, eval=TRUE}
#
##
### Expression level boxplots
##
#

## Create the lists to save data and boxplots
de_data <- list()
de_boxplots <- list()

### Mart object required by format data functions

  # NOTE: Get db data from archived version because useEnsembl was not working (12 Out 2023): 
  # Error in curl::curl_fetch_memory(url, handle = handle) : 
  # SSL peer certificate or SSH remote key was not OK: [uswest.ensembl.org:443] SSL certificate problem: certificate has expired

# ensembl <- useEnsembl(biomart = "ensembl", dataset = "hsapiens_gene_ensembl")
ensembl <- useEnsembl(biomart = "ensembl", dataset = "hsapiens_gene_ensembl", host = "https://jul2023.archive.ensembl.org")


### Mitocarta Gene IDs required by mito data format function
# Get all mitocarta ensembl ids (including the ones with several ids separated by pipe)
mitocarta_gene_ids <-
  unlist(strsplit(
      mitocarta::mitocarta_data$A_Human_MitoCarta3$EnsemblGeneID_mapping_version_20200130,
      "\\|"))


### Create the required directories to save the plots
if (!dir.exists(here('analysis_output/all_genes_boxplots/'))) {
  dir.create(here('analysis_output/all_genes_boxplots/'), recursive = TRUE)}

if (!dir.exists(here('analysis_output/mito_genes_boxplots/'))) {
  dir.create(here('analysis_output/mito_genes_boxplots/'), recursive = TRUE)}

if (!dir.exists(here('analysis_output/dataframes/'))) {
  dir.create(here('analysis_output/dataframes/'), recursive = TRUE)}

```

## Boxplots for Combined with Interaction Dataset | The data discussed in the thesis

```{r boxplots_comb_with, dependson=c("boxplot_funct","load_data","boxplots_prep"), eval=FALSE}

#
##
### Combined with Interaction
##
#


### 1. Format the data for DEG files for each dataset
de_data$combined_de_with_interaction$all <- 
  format_data(diff_exprs_list = combined_de_with_interaction,
                    n_degs_to_plot = 12,
                    mart = ensembl)

de_data$combined_de_with_interaction$mito <- 
  format_data_mito(diff_exprs_list = combined_de_with_interaction,
                    n_degs_to_plot = 12,
                    mart = ensembl,
                    mitocarta_genes = mitocarta_gene_ids)

### 2. Plot the formatted data | ALL GENES 

  ### 2.1. Detailed per organ
de_boxplots$combined_de_with_interaction$all$p1_up_reg <-
  de_data$combined_de_with_interaction$all$extreme_data %>%
  filter(fc_dir == "up") %>%
  plot_organ_and_cancer(counts_data = .,
                                  my_title = "Up regulated DEGs | Combined with Interaction")

de_boxplots$combined_de_with_interaction$all$p2_down_reg <-
  de_data$combined_de_with_interaction$all$extreme_data %>%
  filter(fc_dir == "down") %>%
  plot_organ_and_cancer(counts_data = .,
                                  my_title = "Down regulated DEGs | Combined with Interaction")


  ### 2.2. Per factor only (Metabolic rate and Cancer status)
de_boxplots$combined_de_with_interaction$all$p3_up_reg <-
  de_data$combined_de_with_interaction$all$extreme_data %>%
  filter(fc_dir == "up") %>%
  plot_metab_and_cancer(counts_data = .,
                                  my_title = "Up regulated DEGs | Combined with Interaction")

de_boxplots$combined_de_with_interaction$all$p4_down_reg <-
  de_data$combined_de_with_interaction$all$extreme_data %>%
  filter(fc_dir == "down") %>%
  plot_metab_and_cancer(counts_data = .,
                                  my_title = "Down regulated DEGs | Combined with Interaction")



### 3. Plot the formatted data | MITOCHONDRIAL GENES

  ### 3.1. Detailed per organ
de_boxplots$combined_de_with_interaction$mito$p1_up_reg <-
  de_data$combined_de_with_interaction$mito$extreme_data %>%
  filter(fc_dir == "up") %>%
  plot_organ_and_cancer(counts_data = .,
                                  my_title = "Up regulated Mito DEGs | Combined with Interaction")

de_boxplots$combined_de_with_interaction$mito$p2_down_reg <-
  de_data$combined_de_with_interaction$mito$extreme_data %>%
  filter(fc_dir == "down") %>%
  plot_organ_and_cancer(counts_data = .,
                                  my_title = "Down regulated Mito DEGs | Combined with Interaction")


  ### 3.2. Per factor only (Metabolic rate and Cancer status)

de_boxplots$combined_de_with_interaction$mito$p3_up_reg <-
  de_data$combined_de_with_interaction$mito$extreme_data %>%
  filter(fc_dir == "up") %>%
  plot_metab_and_cancer(counts_data = .,
                                  my_title = "Up regulated Mito DEGs | Combined with Interaction")

de_boxplots$combined_de_with_interaction$mito$p4_down_reg <-
  de_data$combined_de_with_interaction$mito$extreme_data %>%
  filter(fc_dir == "down") %>%
  plot_metab_and_cancer(counts_data = .,
                                  my_title = "Down regulated Mito DEGs | Combined with Interaction")


### 4. Export the boxplots to png

ggsave(file=here("analysis_output/all_genes_boxplots/comb_WITH_interac_up_degs_boxplot.png"),
    plot = de_boxplots$combined_de_with_interaction$all$p1_up_reg, 
    device="png", width = 29.7, height = 21, units = "cm")

ggsave(file=here("analysis_output/all_genes_boxplots/comb_WITH_interac_down_degs_boxplot.png"),
    plot = de_boxplots$combined_de_with_interaction$all$p2_down_reg, 
    device="png", width = 29.7, height = 21, units = "cm")

ggsave(file=here("analysis_output/all_genes_boxplots/comb_WITH_interac_up_degs_boxplot2.png"),
    plot = de_boxplots$combined_de_with_interaction$all$p3_up_reg, 
    device="png", width = 29.7, height = 21, units = "cm")

ggsave(file=here("analysis_output/all_genes_boxplots/comb_WITH_interac_down_degs_boxplot2.png"),
    plot = de_boxplots$combined_de_with_interaction$all$p4_down_reg, 
    device="png", width = 29.7, height = 21, units = "cm")


#### 5. Export mitochondrial boxplots to png

ggsave(file=here("analysis_output/mito_genes_boxplots/comb_WITH_interac_mito_up_degs_boxplot.png"),
    plot = de_boxplots$combined_de_with_interaction$mito$p1_up_reg, 
    device="png", width = 29.7, height = 21, units = "cm")

ggsave(file=here("analysis_output/mito_genes_boxplots/comb_WITH_interac_mito_down_degs_boxplot.png"),
    plot = de_boxplots$combined_de_with_interaction$mito$p2_down_reg, 
    device="png", width = 29.7, height = 21, units = "cm")

ggsave(file=here("analysis_output/mito_genes_boxplots/comb_WITH_interac_mito_up_degs_boxplot2.png"),
    plot = de_boxplots$combined_de_with_interaction$mito$p3_up_reg, 
    device="png", width = 29.7, height = 21, units = "cm")

ggsave(file=here("analysis_output/mito_genes_boxplots/comb_WITH_interac_mito_down_degs_boxplot2.png"),
    plot = de_boxplots$combined_de_with_interaction$mito$p4_down_reg, 
    device="png", width = 29.7, height = 21, units = "cm")


#### 6. Save to file the summary tables
write.table(de_data$combined_de_with_interaction$all$sum_table,
            file = here("analysis_output/dataframes/up_down_reg_boxplot_all_df.csv"),
            quote = TRUE, sep = "\t", row.names = FALSE, col.names = TRUE)

write.table(de_data$combined_de_with_interaction$mito$sum_table,
            file = here("analysis_output/dataframes/up_down_reg_boxplot_mito_df.csv"),
            quote = TRUE, sep = "\t", row.names = FALSE, col.names = TRUE)

```


## Boxplots for remaining datasets

```{r boxplots_other_datasets, dependson="boxplots_prep", eval=FALSE}

#
##
### Format the data for DEG files for each dataset
##
#

#### GTEx ###
de_data$gtex$all <- 
  format_data(diff_exprs_list = gtex_de,
                    n_degs_to_plot = 10,
                    mart = ensembl)
de_data$gtex$mito <- 
  format_data_mito(diff_exprs_list = gtex_de,
                    n_degs_to_plot = 10,
                    mart = ensembl,
                    mitocarta_genes = mitocarta_gene_ids)

#### TCGA ###
de_data$tcga$all <- 
  format_data(diff_exprs_list = tcga_de,
                    n_degs_to_plot = 10,
                    mart = ensembl)
de_data$tcga$mito <- 
  format_data_mito(diff_exprs_list = tcga_de,
                    n_degs_to_plot = 10,
                    mart = ensembl,
                    mitocarta_genes = mitocarta_gene_ids)

#### Combined no interaction ###
de_data$comb_no$all <- 
  format_data(diff_exprs_list = combined_de_no_interaction,
                    n_degs_to_plot = 10,
                    mart = ensembl)
de_data$comb_no$mito <- 
  format_data_mito(diff_exprs_list = combined_de_no_interaction,
                    n_degs_to_plot = 10,
                    mart = ensembl,
                    mitocarta_genes = mitocarta_gene_ids)

#### Combined only interaction ###
de_data$comb_only$all <- 
  format_data(diff_exprs_list = combined_de_only_interaction,
                    n_degs_to_plot = 10,
                    mart = ensembl)
de_data$comb_only$mito <- 
  format_data_mito(diff_exprs_list = combined_de_only_interaction,
                    n_degs_to_plot = 10,
                    mart = ensembl,
                    mitocarta_genes = mitocarta_gene_ids)


#
##
### Plot the formatted data | ALL GENES
##
#

#### Combined no interaction ###
de_boxplots$comb_no$all$p1_extreme <-
  plot_organ_and_cancer(counts_data = de_data$comb_no$all$extreme_data,
                                  my_title = "Extreme DEGs | Combined no Interaction")

#### Combined only interaction ###
de_boxplots$comb_only$all$p1_extreme <-
  plot_organ_and_cancer(counts_data = de_data$comb_only$all$extreme_data,
                                  my_title = "Extreme DEGs | Combined only Interaction")

# #### GTEx ### Not working! Error: At least one layer must contain all faceting variables: `gene_id`
# de_boxplots$gtex$all$p1_extreme <-
#   plot_organ(counts_data = de_data$gtex_de$all$extreme_data,
#                                   my_title = "Extreme DEGs | GTEx")
# 
# #### TCGA ### Not working! Error: At least one layer must contain all faceting variables: `gene_id`
# de_boxplots$tcga$all$p1_extreme <-
#   plot_organ(counts_data = de_data$tcga_de$all$extreme_data,
#                                   my_title = "Extreme DEGs | TCGA")

#
##
### Plot the formatted data | MITOCHONDRIAL GENES
##
#

#### Combined no interaction ###
de_boxplots$comb_no$mito$p1_extreme <-
  plot_organ_and_cancer(counts_data = de_data$comb_no$mito$extreme_data,
                                  my_title = "Extreme Mito DEGs | Combined no Interaction")

#### Combined only interaction ###
de_boxplots$comb_only$mito$p1_extreme <-
  plot_organ_and_cancer(counts_data = de_data$comb_only$mito$extreme_data,
                                  my_title = "Extreme Mito DEGs | Combined only Interaction")

# #### GTEx ### Not working!
# de_boxplots$gtex$mito$p1_extreme <-
#   plot_organ(counts_data = de_data$gtex_de$mito$extreme_data,
#                                   my_title = "Extreme Mito DEGs | GTEx")
# 
# #### TCGA ### Not working!
# de_boxplots$tcga$mito$p1_extreme <-
#   plot_organ(counts_data = de_data$tcga_de$mito$extreme_data,
#                                   my_title = "Extreme Mito DEGs | TCGA")

#
##
### Export the boxplots to png
##
#

#### Boxplots all genes ###

# ### GTEx ###
# ggsave(file=here("analysis_output/all_genes_boxplots/gtex_degs_boxplot.png"),
#     plot = de_boxplots$gtex$all$p1_extreme, 
#     device="png", width = 29.7, height = 21, units = "cm")
# 
# ### TCGA ###
# ggsave(file=here("analysis_output/all_genes_boxplots/tcga_degs_boxplot.png"),
#     plot = de_boxplots$tcga$all$p1_extreme, 
#     device="png", width = 29.7, height = 21, units = "cm")

### Combined no interaction ###
ggsave(file=here("analysis_output/all_genes_boxplots/comb_no_interac_degs_boxplot.png"),
    plot = de_boxplots$comb_no$all$p1_extreme, 
    device="png", width = 29.7, height = 21, units = "cm")

### Combined only interaction ###
ggsave(file=here("analysis_output/all_genes_boxplots/comb_only_interac_degs_boxplot.png"),
    plot = de_boxplots$comb_only$all$p1_extreme, 
    device="png", width = 29.7, height = 21, units = "cm")


#### Boxplots mitochondrial genes ###

# ### GTEx ### 
# ggsave(file=here("analysis_output/mito_genes_boxplots/gtex_mito_degs_boxplot.png"),
#     plot = de_boxplots$gtex$mito$p1_extreme, 
#     device="png", width = 29.7, height = 21, units = "cm")
# 
# ### TCGA ### 
# ggsave(file=here("analysis_output/mito_genes_boxplots/tcga_mito_degs_boxplot.png"),
#     plot = de_boxplots$tcga$mito$p1_extreme, 
#     device="png", width = 29.7, height = 21, units = "cm")

### Combined no interaction ###
ggsave(file=here("analysis_output/mito_genes_boxplots/comb_no_interac_mito_degs_boxplot.png"),
    plot = de_boxplots$comb_no$mito$p1_extreme, 
    device="png", width = 29.7, height = 21, units = "cm")

### Combined only interaction ###
ggsave(file=here("analysis_output/mito_genes_boxplots/comb_only_interac_mito_degs_boxplot.png"),
    plot = de_boxplots$comb_only$mito$p1_extreme, 
    device="png", width = 29.7, height = 21, units = "cm")


```

## Create tables for Coefficients of the different datasets and linear models | For future comparisons between models

```{r coef, dependson="load_data", eval=TRUE}
#
##
### Coefficients | Merge data to compare
##
#

# Create list
lrt_coef_df <- list()
lrt_coef_df$separate <- list()
lrt_coef_df$merged <- list()

# Get coefficients in data frame format and change the column names to include the dataset name (except the gene_id column)
lrt_coef_df$separate$gtex <- 
  data.frame(gene_id = rownames(gtex_de$lrt$coefficients), 
             gtex_de$lrt$coefficients, row.names = NULL)
colnames (lrt_coef_df$separate$gtex)[-1] <- paste0(colnames(lrt_coef_df$separate$gtex)[-1], "_gtex")

lrt_coef_df$separate$tcga <- 
  data.frame(gene_id = rownames(tcga_de$lrt$coefficients), 
             tcga_de$lrt$coefficients, row.names = NULL)
colnames (lrt_coef_df$separate$tcga)[-1] <- paste0(colnames(lrt_coef_df$separate$tcga)[-1], "_tcga")

lrt_coef_df$separate$combined_no_interaction <- 
  data.frame(gene_id = rownames(combined_de_no_interaction$lrt$coefficients), 
             combined_de_no_interaction$lrt$coefficients, row.names = NULL)
colnames (lrt_coef_df$separate$combined_no_interaction)[-1] <- paste0(colnames(lrt_coef_df$separate$combined_no_interaction)[-1], "_comb_NO")

lrt_coef_df$separate$combined_only_interaction <- 
  data.frame(gene_id = rownames(combined_de_only_interaction$lrt$coefficients), 
             combined_de_only_interaction$lrt$coefficients, row.names = NULL)
colnames (lrt_coef_df$separate$combined_only_interaction)[-1] <- paste0(colnames(lrt_coef_df$separate$combined_only_interaction)[-1], "_comb_ONLY")

lrt_coef_df$separate$combined_with_interaction <- 
  data.frame(gene_id = rownames(combined_de_with_interaction$lrt$coefficients), 
             combined_de_with_interaction$lrt$coefficients, row.names = NULL)
colnames (lrt_coef_df$separate$combined_with_interaction)[-1] <- paste0(colnames(lrt_coef_df$separate$combined_with_interaction)[-1], "_comb_WITH")


# Merge all dataframes by common gene ids
lrt_coef_df$merged$wide <- Reduce(function(x,y) inner_join (x,y, by="gene_id"), lrt_coef_df$separate)


# Pivot the dataframe to make the table tidy
lrt_coef_df$merged$long <- lrt_coef_df$merged$wide %>%
  pivot_longer(cols = !gene_id,
               cols_vary = "slowest",
               names_to = "coeff_name",
               values_to = "coeff_value") %>%
  mutate(coeff_name = factor(coeff_name,
                                levels = colnames(lrt_coef_df$merged$wide)[-1]))

## Save the coefficient tables
saveRDS(lrt_coef_df$merged, file = here("data/analysis/lrt_coef_df_merged.RDS"))

```

## Create tables of LFC and CPMs of the different datasets and linear models | For future comparisons between models

```{r top_genes, dependson="load_data", eval=TRUE}

#
##
### Top genes | Merge data to compare
##
#

# Create list
top_genes_df <- list()
top_genes_df$separate <- list()
top_genes_df$merged <- list()

# Get top genes table and change the column names to include the dataset name (except the gene_id column)
top_genes_df$separate$gtex <- 
  data.frame(gene_id = rownames(gtex_de$topGenes$table), 
             gtex_de$topGenes$table, row.names = NULL)
colnames (top_genes_df$separate$gtex)[-1] <- paste0(colnames(top_genes_df$separate$gtex)[-1], "_gtex")

top_genes_df$separate$tcga <- 
  data.frame(gene_id = rownames(tcga_de$topGenes$table), 
             tcga_de$topGenes$table, row.names = NULL)
colnames (top_genes_df$separate$tcga)[-1] <- paste0(colnames(top_genes_df$separate$tcga)[-1], "_tcga")

top_genes_df$separate$combined_no_interaction <- 
  data.frame(gene_id = rownames(combined_de_no_interaction$topGenes$table), 
             combined_de_no_interaction$topGenes$table, row.names = NULL)
colnames (top_genes_df$separate$combined_no_interaction)[-1] <- paste0(colnames(top_genes_df$separate$combined_no_interaction)[-1], "_comb_NO")

top_genes_df$separate$combined_only_interaction <- 
  data.frame(gene_id = rownames(combined_de_only_interaction$topGenes$table), 
             combined_de_only_interaction$topGenes$table, row.names = NULL)
colnames (top_genes_df$separate$combined_only_interaction)[-1] <- paste0(colnames(top_genes_df$separate$combined_only_interaction)[-1], "_comb_ONLY")

top_genes_df$separate$combined_with_interaction <- 
  data.frame(gene_id = rownames(combined_de_with_interaction$topGenes$table), 
             combined_de_with_interaction$topGenes$table, row.names = NULL)
colnames (top_genes_df$separate$combined_with_interaction)[-1] <- paste0(colnames(top_genes_df$separate$combined_with_interaction)[-1], "_comb_WITH")


# Merge all dataframes by common gene ids
top_genes_df$merged$wide <- Reduce(function(x,y) inner_join (x,y, by="gene_id"), top_genes_df$separate)


# Pivot LogFC and LogCPM data to make it tidy

top_genes_df$merged$logFC_long <-
  top_genes_df$merged$wide %>%
  dplyr::select(gene_id, contains("LogFC")) %>%
  pivot_longer(cols = !gene_id, 
               names_to = "dataset_model", 
               values_to = "LogFC",
               cols_vary = "fastest") %>%
  mutate(dataset_model = factor(dataset_model,
                                levels = c("logFC_gtex","logFC_tcga",
                                           "logFC_comb_NO",
                                           "logFC_comb_ONLY",
                                           "logFC_comb_WITH") ) )
  
top_genes_df$merged$logCPM_long <-
  top_genes_df$merged$wide %>%
  dplyr::select(gene_id, contains("LogCPM")) %>%
  pivot_longer(cols = !gene_id, 
               names_to = "dataset_model", 
               values_to = "LogCPM",
               cols_vary = "fastest") %>%
  mutate(dataset_model = factor(dataset_model,
                                levels = c("logCPM_gtex","logCPM_tcga",
                                           "logCPM_comb_NO",
                                           "logCPM_comb_ONLY",
                                           "logCPM_comb_WITH") ) )

top_genes_df$merged$FDR_long <-
  top_genes_df$merged$wide %>%
  dplyr::select(gene_id, contains("FDR")) %>%
  pivot_longer(cols = !gene_id, 
               names_to = "dataset_model", 
               values_to = "FDR",
               cols_vary = "fastest") %>%
  mutate(dataset_model = factor(dataset_model,
                                levels = c("FDR_gtex","FDR_tcga",
                                           "FDR_comb_NO",
                                           "FDR_comb_ONLY",
                                           "FDR_comb_WITH") ) )

## Save the top genes tables
saveRDS(top_genes_df$merged, file = here("data/analysis/top_genes_df_merged.RDS"))

```
