==> code/4.functional_enrichment.Rmd <==
---
title: "MitoSignatures"
author: "Isabel Duarte"
date: "2023-05-31"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=TRUE, message=FALSE, warnings=FALSE, cache = TRUE)

library(here)
library(tidyverse)
library(edgeR)
library(AnnotationDbi)
library(org.Hs.eg.db)
library(gprofiler2)

```

## Load Differential Expression analyses results 

```{r load_data, eval=FALSE}

# Load data
gtex_de <- readRDS(here("data/analysis/gtex_de.RDS"))
tcga_de <- readRDS(here("data/analysis/tcga_de.RDS"))
combined_de_with_interaction <- readRDS(here("data/analysis/combined_de_with_interaction.RDS"))


```

## Functional enrichment analysis | GTEx

```{r gtex_func_enrich, eval=TRUE}

#  Create a list to hold the functional enrichment objects
gtex_fun <- list()

# Get Entrez Gene identifiers (required by edgeR functions)
rownames(gtex_de$lrt) %>%
  gsub('\\.\\d+', '', .) -> gtex_fun$ensembl_ids

# Convert Ensembl IDs to Entrez Gene IDs
columns(org.Hs.eg.db)

gtex_fun$entrez_ids <- mapIds(org.Hs.eg.db, 
                              keys = gtex_fun$ensembl_ids, 
                              column = "ENTREZID", 
                              keytype = "ENSEMBL", 
                              multiVals = "first")

# GO analysis
gtex_fun$go <- edgeR::goana.DGELRT(gtex_de$lrt, 
                                   geneid = gtex_fun$entrez_ids,
                                   species = "Hs")
gtex_fun$top_go <- topGO(gtex_fun$go, sort="up")


# KEGG analysis # FAILS in 7 Oct 2023 | Error with https://rest.kegg.jp/link/pathway/hsa
# gtex_fun$kegg <- edgeR::kegga.DGELRT(gtex_de$lrt, 
#                                      geneid = gtex_fun$entrez_ids,
#                                      species = "Hs")
# gtex_fun$top_kegg <- topKEGG(gtex_fun$kegg, sort="up")

# Print a sentence confirming that this chunk finished computing
cat("Finished the GTEx functional enrichment analysis.")

```

## Functional enrichment analysis | TCGA

```{r tcga_func_enrich, eval=TRUE}

#  Create a list to hold the functional enrichment objects
tcga_fun <- list()

# Get Entrez Gene identifiers (required by edgeR functions)
rownames(tcga_de$lrt) %>%
  gsub('\\.\\d+', '', .) -> tcga_fun$ensembl_ids

# Convert Ensembl IDs to Entrez Gene IDs
tcga_fun$entrez_ids <- mapIds(org.Hs.eg.db, 
                              keys = tcga_fun$ensembl_ids, 
                              column = "ENTREZID", 
                              keytype = "ENSEMBL",
                              multiVals = "first")

# GO analysis
tcga_fun$go <- edgeR::goana.DGELRT(tcga_de$lrt, 
                                   geneid = tcga_fun$entrez_ids,
                                   species = "Hs")
tcga_fun$top_go <- topGO(tcga_fun$go, sort="up")


# KEGG analysis # FAILS in 7 Oct 2023 | Error with https://rest.kegg.jp/link/pathway/hsa
# tcga_fun$kegg <- edgeR::kegga.DGELRT(tcga_de$lrt, 
#                                      geneid = tcga_fun$entrez_ids,
#                                      species = "Hs")
# tcga_fun$top_kegg <- topKEGG(tcga_fun$kegg, sort="up")

# Print a sentence confirming that this chunk finished computing
cat("Finished the TCGA functional enrichment analysis.")

```

## Functional enrichment analysis | Combined GTEx TCGA

```{r combined_func_enrich, eval=TRUE}

#  Create a list to hold the functional enrichment objects
combined_with_inter_fun <- list()

# Get Entrez Gene identifiers (required by edgeR functions)
rownames(combined_de_with_interaction$lrt) %>%
  gsub('\\.\\d+', '', .) -> combined_with_inter_fun$ensembl_ids

# Convert Ensembl IDs to Entrez Gene IDs
combined_with_inter_fun$entrez_ids <- mapIds(org.Hs.eg.db, 
                              keys = combined_with_inter_fun$ensembl_ids, 
                              column = "ENTREZID", 
                              keytype = "ENSEMBL",
                              multiVals = "first")

# GO analysis
combined_with_inter_fun$go <- edgeR::goana.DGELRT(combined_de_with_interaction$lrt, 
                                   geneid = combined_with_inter_fun$entrez_ids,
                                   species = "Hs")
combined_with_inter_fun$top_go <- topGO(combined_with_inter_fun$go, sort="up")


# KEGG analysis # FAILS in 7 Oct 2023 | Error with https://rest.kegg.jp/link/pathway/hsa
# combined_with_inter_fun$kegg <- edgeR::kegga.DGELRT(combined_de_with_interaction$lrt, 
#                                      geneid = combined_with_inter_fun$entrez_ids,
#                                      species = "Hs")
# combined_with_inter_fun$top_kegg <- topKEGG(combined_with_inter_fun$kegg, sort="up")

# Print a sentence confirming that this chunk finished computing
cat("Finished the Combined GTEx TCGA functional enrichment analysis.")

```


## Functional enrichment | Using gprofiler2

```{r funct_enrich_gprofiler, eval=TRUE}
#
##
### GTEx | Perform functional enrichment analysis
##
#
gtex_fun$gprofiler <- gprofiler2::gost(
  query = gsub('\\.\\d+', '', rownames(gtex_de$topGenes$table)[1:500]),
  organism = "hsapiens",
  ordered_query = FALSE,
  user_threshold = 0.05,
  correction_method = "fdr",
  significant = TRUE
)

# Plot the results
gprofiler2::gostplot(gtex_fun$gprofiler)

# Print a sentence confirming that this chunk finished computing
cat("Finished the gprofiler GTEx functional enrichment analysis.")

#
##
### TCGA | Perform functional enrichment analysis
##
#
tcga_fun$gprofiler <- gprofiler2::gost(
  query = gsub('\\.\\d+', '', rownames(tcga_de$topGenes$table)[1:500]),
  organism = "hsapiens",
  ordered_query = FALSE,
  user_threshold = 0.05,
  correction_method = "fdr",
  significant = TRUE
)

# Plot the results
gprofiler2::gostplot(tcga_fun$gprofiler)

# Print a sentence confirming that this chunk finished computing
cat("Finished the gprofiler TCGA functional enrichment analysis.")

#
##
### Combined GTEx TCGA | Perform functional enrichment analysis
##
#
combined_with_inter_fun$gprofiler <- gprofiler2::gost(
  query = gsub('\\.\\d+', '', rownames(combined_de_with_interaction$topGenes$table)[1:500]),
  organism = "hsapiens",
  ordered_query = FALSE,
  user_threshold = 0.05,
  correction_method = "fdr",
  significant = TRUE
)

# Plot the results
combined_with_inter_fun$gprofiler$plot<- gprofiler2::gostplot(combined_with_inter_fun$gprofiler,
                                                              interactive = FALSE)
gprofiler2::publish_gostplot(combined_with_inter_fun$gprofiler$plot,
                             filename = here("analysis_output/comb_with_inter_gostplot.pdf"),
                             width = 10, height = 6)

# Print a sentence confirming that this chunk finished computing
cat("Finished the gprofiler Combined functional enrichment analysis.")

```

## Save the functional enrichment results

```{r save_fun_results, eval=TRUE}

# Save the gtex functional analysis results
saveRDS(gtex_fun, file = here("data/analysis/gtex_fun.RDS"))

# Save the tcga functional analysis results
saveRDS(tcga_fun, file = here("data/analysis/tcga_fun.RDS"))

# Save the gtex functional analysis results
saveRDS(combined_with_inter_fun, file = here("data/analysis/combined_with_inter_fun.RDS"))

# Print a sentence confirming that this chunk finished computing
cat("Finished saving to file the functional enrichment analysis results.")

```
