==> code/5.clustering.Rmd <==
---
title: "MitoSignatures"
author: "Catarina Ferreira & Isabel Duarte"
date: "2023-07-17"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo=TRUE, message=FALSE, warnings=FALSE, cache = FALSE)

library(here)
library(tidyverse)
library(mitocarta)
library(biomaRt)
library(edgeR)
library(RColorBrewer)

```

## Load and Format data

```{r load_data}

## Read the file with the diff expression data
combined_de_with_interaction <- readRDS(file = here("data/analysis/combined_de_with_interaction.RDS"))

# Mitocarta Gene IDs
mitocarta_gene_ids <-
  unlist(
    strsplit(
      mitocarta::mitocarta_data$A_Human_MitoCarta3$EnsemblGeneID_mapping_version_20200130,
      "\\|"))

## Get the index positions from the dgelist object that contain the mitochondrial genes 
mito_genes_index <- gsub("\\..*$", "", 
                         row.names(combined_de_with_interaction$dge_data_filtered)) %in% 
  mitocarta_gene_ids

## Get a new DGELIST object only with mitochondrial genes 
mito_dge_data_filtered <- combined_de_with_interaction$dge_data_filtered[mito_genes_index, ]

## Get the normalized CPM values for the mitochondrial genes
dgelist_log2cpm_mito <- cpm(mito_dge_data_filtered,
                       normalized.lib.sizes = TRUE,
                       log = TRUE, prior.count = 2)


## Metadata | Create vector of colors for samples
combined_de_with_interaction$data_clean$metadata %>%
  filter(sample_id  %in% colnames(dgelist_log2cpm_mito)) %>%
  mutate(sample_type = paste(cancer_status, organ, sep = "_")) %>%
  mutate(sample_col = case_when(sample_type == "cancer_bladder" ~ "tomato1",
                                sample_type == "non_cancer_bladder" ~ "tomato3",
                                sample_type == "cancer_skin" ~ "coral3",
                                sample_type == "non_cancer_skin" ~ "coral",
                                sample_type == "cancer_colon" ~ "green4",
                                sample_type == "non_cancer_colon" ~ "limegreen",
                                sample_type == "cancer_brain" ~ "deeppink",
                                sample_type == "non_cancer_brain" ~ "deeppink3",
                                sample_type == "cancer_liver" ~ "mediumpurple3",
                                sample_type == "non_cancer_liver" ~ "mediumpurple1",
                                sample_type == "cancer_kidney" ~ "darkgoldenrod1",
                                sample_type == "non_cancer_kidney" ~ "darkgoldenrod3")) %>%
  mutate(labels = case_when(sample_type == "cancer_bladder" ~ "Bladder Cancer",
                                sample_type == "non_cancer_bladder" ~ "Bladded Non-cancer",
                                sample_type == "cancer_skin" ~ "Skin Cancer",
                                sample_type == "non_cancer_skin" ~ "Skin Non-cancer",
                                sample_type == "cancer_colon" ~ "Colon Cancer",
                                sample_type == "non_cancer_colon" ~ "Colon Non-cancer",
                                sample_type == "cancer_brain" ~ "Brain Cancer",
                                sample_type == "non_cancer_brain" ~ "Brain Non-cancer",
                                sample_type == "cancer_liver" ~ "Liver Cancer",
                                sample_type == "non_cancer_liver" ~ "Liver Non-cancer",
                                sample_type == "cancer_kidney" ~ "Kidney Cancer",
                                sample_type == "non_cancer_kidney" ~ "Kidney Non-cancer")) %>%
  dplyr::arrange(organ,cancer_status) -> metadata_with_colors_ordered


## Order the columns of the original expression matrix by sample type
   ## (to keep all sample types together in the heatmap)
dgelist_log2cpm_mito_ordered <- dgelist_log2cpm_mito[ , 
                                           match(metadata_with_colors_ordered$sample_id,
                                                 colnames(dgelist_log2cpm_mito))
                                           ]
# Clean up
rm(dgelist_log2cpm_mito)

```

## Hierarchical clustering

```{r hierarchical_clustering}

set.seed(42)

# Change the count matrix row names to gene symbols

  # Get the ensembl gene annotations
  ensembl <- useEnsembl(biomart = "ensembl", dataset = "hsapiens_gene_ensembl")
  
  # Fetch gene symbols and gene names
  gene_info <- getBM(
    attributes = c("ensembl_gene_id", "external_gene_name", "description"),
    filters = "ensembl_gene_id",
    values = mitocarta_gene_ids,
    mart = ensembl
  )
  
  # Change row names from count matrix to gene symbols
  row.names(dgelist_log2cpm_mito_ordered) <- gene_info$external_gene_name[gene_info$ensembl_gene_id %in% gsub("\\..*$", "", row.names(dgelist_log2cpm_mito_ordered))]
  
  
 ## Calculate the distance matrix - SLOW STEP - UNCOMMENT WHEN NEEDED
mito_dist_matrix <- dist(dgelist_log2cpm_mito_ordered, method = "manhattan")

## Save the distance matrix object
saveRDS(mito_dist_matrix, file = here("data/analysis/clustering_mito_distance_matrix.RDS"))

## DIFFERENT TYPES OF HIERARCHICAL CLUSTERING
## Single link
hclust_min <- hclust(mito_dist_matrix, method = "single")

## Complete link
hclust_max <- hclust(mito_dist_matrix, method = "complete")

## Group average
hclust_group_av <- hclust(mito_dist_matrix, method = "average")


### Visualize the dendrogram
plot(hclust_min, cex=0.4, hang = -1,
     main = "Dendrogram of Hierarchical Clustering - Single linkage")

plot(hclust_max, cex=0.4, hang = -1, 
     main = "Dendrogram of Hierarchical Clustering - Complete linkage")

plot(hclust_group_av, cex=0.5, hang = -1, 
     main = "Dendrogram of Hierarchical Clustering - Group average")
rect.hclust(hclust_group_av, k = 5, border = rainbow(5))


### Save the hierarchical clustering dendrograms
  if (!dir.exists(here('analysis_output/heatmaps/'))) {
    dir.create(here('analysis_output/heatmaps/'), recursive = TRUE)}

pdf(file=here("analysis_output/heatmaps/hierarchical_single.pdf"),
      width = 9, height = 6)
    plot(hclust_min, cex=0.4, hang=-1,
         ylab="Distance", xlab="Gene Symbol (mitochondrial proteins)",
         main = "Hierarchical Clustering - Single linkage")
dev.off()

pdf(file=here("analysis_output/heatmaps/hierarchical_complete.pdf"),
      width = 9, height = 6)
    plot(hclust_max, cex=0.4, hang=-1,
         ylab="Distance", xlab="Gene Symbol (mitochondrial proteins)",
         main = "Hierarchical Clustering - Complete linkage")
dev.off()

pdf(file=here("analysis_output/heatmaps/hierarchical_group.pdf"),
      width = 9, height = 6)
    plot(hclust_group_av, cex=0.4, hang=-1, 
         ylab="Distance", xlab="Gene Symbol (mitochondrial proteins)",
         main = "Hierarchical Clustering - Group average")
    rect.hclust(hclust_group_av, k = 5, border = rainbow(5))
dev.off()

###


#### Heatmap visualization

# Change colnames from count matrix 
dgelist_log2cpm_mito_ordered_2 <- dgelist_log2cpm_mito_ordered
colnames(dgelist_log2cpm_mito_ordered_2) <- metadata_with_colors_ordered$labels

# Save the heatmap as a PDF file
png((file=here("analysis_output/heatmaps/heatmap_with_interaction.png")), 
    width = 2480, height = 1754, units = "px")  

  heatmap(dgelist_log2cpm_mito_ordered_2, 
          Rowv = order(hclust_group_av$order), 
          Colv = NA,
          scale = "none", 
          symm = FALSE, 
          cexRow=2, cexCol=0.1, margins = c(10, 10),
          # main="Heatmap | Group linkage Hierarchical clustering",
          col= colorRampPalette(brewer.pal(8, "Blues"))(25),
          ColSideColors = metadata_with_colors_ordered$sample_col)
 
  # Column color labels
  mtext( "Bladder", side=3, cex = 2.5, line=0, at=c(0.25))
  mtext( "NC", side=3, cex = 2, line=-1.8, at=c(0.235))
  mtext( "C", side=3, cex = 2, line=-1.8, at=c(0.26))
  
  mtext( "Colon", side=3, cex = 2.5, line=0, at=c(0.35))
  mtext( "NC", side=3, cex = 2, line=-1.8, at=c(0.32))
  mtext( "C", side=3, cex = 2, line=-1.8, at=c(0.38))
  
  mtext( "Skin", side=3, cex = 2.5, line=0, at=c(0.51))
  mtext( "NC", side=3, cex = 2, line=-1.8, at=c(0.48))
  mtext( "C", side=3, cex = 2, line=-1.8, at=c(0.57))
  
  mtext( "Brain", side=3, cex = 2.5, line=0, at=c(0.64))
  mtext( "NC", side=3, cex = 2, line=-1.8, at=c(0.615))
  mtext( "C", side=3, cex = 2, line=-1.8, at=c(0.66))
  
  mtext( "Kidney", side=3, cex = 2.5, line=0, at=c(0.76))
  mtext( "NC", side=3, cex = 2, line=-1.8, at=c(0.705))
  mtext( "C", side=3, cex = 2, line=-1.8, at=c(0.775))
  
  mtext( "Liver", side=3, cex = 2.5, line=0, at=c(0.85))
  mtext( "NC", side=3, cex = 2, line=-1.8, at=c(0.825))
  mtext( "C", side=3, cex = 2, line=-1.8, at=c(0.86))
  
dev.off()

```


## Soft clustering | Fuzzy clustering with cmeans

```{r fuzzy_clustering, eval=TRUE}

set.seed(3)

## Run cmeans clustering algorithm, with 20 clusters, and 1.25 fuzzification parameter m
cmeans_res <- e1071::cmeans(dgelist_log2cpm_mito_ordered, 
                            centers=20, m=1.25, 
                            dist = "manhattan", method = "cmeans")

## Save result
write_rds(cmeans_res, file = here("data/analysis/cmeans_result.RDS"))

#
##
### Plot clusters | Adapted from plotting function in Mfuzz package
##
#

my_mfuzz.plot2 <- function (my_matrix, cl, mfrow = c(1, 1), colo, min.mem = 0, time.labels, 
    time.points, ylim.set = c(0, 0), xlab = "Time", ylab = "Expression changes", 
    x11 = TRUE, ax.col = "black", bg = "white", col.axis = "black", 
    col.lab = "black", col.main = "black", col.sub = "black", 
    col = "black", centre = FALSE, centre.col = "black", centre.lwd = 2, 
    Xwidth = 5, Xheight = 5, single = FALSE, ...) {
  
    clusterindex <- cl[[3]]
    memship <- cl[[4]]
    memship[memship < min.mem] <- -1
    colorindex <- integer(dim(my_matrix)[[1]])
    if (missing(colo)) {
        colo <- c("#FF0000", "#FF1800", "#FF3000", "#FF4800", 
            "#FF6000", "#FF7800", "#FF8F00", "#FFA700", "#FFBF00", 
            "#FFD700", "#FFEF00", "#F7FF00", "#DFFF00", "#C7FF00", 
            "#AFFF00", "#97FF00", "#80FF00", "#68FF00", "#50FF00", 
            "#38FF00", "#20FF00", "#08FF00", "#00FF10", "#00FF28", 
            "#00FF40", "#00FF58", "#00FF70", "#00FF87", "#00FF9F", 
            "#00FFB7", "#00FFCF", "#00FFE7", "#00FFFF", "#00E7FF", 
            "#00CFFF", "#00B7FF", "#009FFF", "#0087FF", "#0070FF", 
            "#0058FF", "#0040FF", "#0028FF", "#0010FF", "#0800FF", 
            "#2000FF", "#3800FF", "#5000FF", "#6800FF", "#8000FF", 
            "#9700FF", "#AF00FF", "#C700FF", "#DF00FF", "#F700FF", 
            "#FF00EF", "#FF00D7", "#FF00BF", "#FF00A7", "#FF008F", 
            "#FF0078", "#FF0060", "#FF0048", "#FF0030", "#FF0018")
    }
    else {
        if (colo == "fancy") {
            fancy.blue <- c(c(255:0), rep(0, length(c(255:0))), 
                rep(0, length(c(255:150))))
            fancy.green <- c(c(0:255), c(255:0), rep(0, length(c(255:150))))
            fancy.red <- c(c(0:255), rep(255, length(c(255:0))), 
                c(255:150))
            colo <- rgb(b = fancy.blue/255, g = fancy.green/255, 
                r = fancy.red/255)
        }
    }
    colorseq <- seq(0, 1, length = length(colo))
    for (j in 1:dim(cl[[1]])[[1]]) {
        if (single) 
            j <- single
        tmp <- my_matrix[clusterindex == j, , drop = FALSE]
        tmpmem <- memship[clusterindex == j, j]
        if (((j - 1)%%(mfrow[1] * mfrow[2])) == 0 | single) {
            if (x11) 
                X11(width = Xwidth, height = Xheight)
            if (sum(clusterindex == j) == 0) {
                ymin <- -1
                ymax <- +1
            }
            else {
                ymin <- min(tmp)
                ymax <- max(tmp)
            }
            if (sum(ylim.set == c(0, 0)) == 2) {
                ylim <- c(ymin, ymax)
            }
            else {
                ylim <- ylim.set
            }
            if (!is.na(sum(mfrow))) {
                par(mfrow = mfrow, bg = bg, col.axis = col.axis, 
                  col.lab = col.lab, col.main = col.main, col.sub = col.sub, 
                  col = col)
            }
            else {
                par(bg = bg, col.axis = col.axis, col.lab = col.lab, 
                  col.main = col.main, col.sub = col.sub, col = col)
            }
            xlim.tmp <- c(1, dim(my_matrix)[[2]])
            if (!(missing(time.points))) 
                xlim.tmp <- c(min(time.points), max(time.points))
            plot.default(x = NA, xlim = xlim.tmp, ylim = ylim, 
                xlab = xlab, ylab = ylab, main = paste("Cluster", 
                  j), axes = FALSE, ...)
            if (missing(time.labels) && missing(time.points)) {
                axis(1, 1:dim(my_matrix)[[2]], c(1:dim(my_matrix)[[2]]), 
                  col = ax.col, ...)
                axis(2, col = ax.col, ...)
            }
            if (missing(time.labels) && !(missing(time.points))) {
                axis(1, time.points, 1:length(time.points), time.points, 
                  col = ax.col, ...)
                axis(2, col = ax.col, ...)
            }
            if (missing(time.points) & !(missing(time.labels))) {
                axis(1, 1:dim(my_matrix)[[2]], time.labels, 
                  col = ax.col, ...)
                axis(2, col = ax.col, ...)
            }
            if (!(missing(time.points)) & !(missing(time.labels))) {
                axis(1, time.points, time.labels, col = ax.col, 
                  ...)
                axis(2, col = ax.col, ...)
            }
        }
        else {
            if (sum(clusterindex == j) == 0) {
                ymin <- -1
                ymax <- +1
            }
            else {
                ymin <- min(tmp)
                ymax <- max(tmp)
            }
            if (sum(ylim.set == c(0, 0)) == 2) {
                ylim <- c(ymin, ymax)
            }
            else {
                ylim <- ylim.set
            }
            xlim.tmp <- c(1, dim(my_matrix)[[2]])
            if (!(missing(time.points))) 
                xlim.tmp <- c(min(time.points), max(time.points))
            plot.default(x = NA, xlim = xlim.tmp, ylim = ylim, 
                xlab = xlab, ylab = ylab, main = paste("Cluster", 
                  j), axes = FALSE, ...)
            if (missing(time.labels) && missing(time.points)) {
                axis(1, 1:dim(my_matrix)[[2]], c(1:dim(my_matrix)[[2]]), 
                  col = ax.col, ...)
                axis(2, col = ax.col, ...)
            }
            if (missing(time.labels) && !(missing(time.points))) {
                axis(1, time.points, 1:length(time.points), time.points, 
                  col = ax.col, ...)
                axis(2, col = ax.col, ...)
            }
            if (missing(time.points) & !(missing(time.labels))) {
                axis(1, 1:dim(my_matrix)[[2]], time.labels, 
                  col = ax.col, ...)
                axis(2, col = ax.col, ...)
            }
            if (!(missing(time.points)) & !(missing(time.labels))) {
                axis(1, time.points, time.labels, col = ax.col, 
                  ...)
                axis(2, col = ax.col, ...)
            }
        }
        if (length(tmpmem) > 0) {
            for (jj in 1:(length(colorseq) - 1)) {
                tmpcol <- (tmpmem >= colorseq[jj] & tmpmem <= 
                  colorseq[jj + 1])
                if (sum(tmpcol) > 0) {
                  tmpind <- which(tmpcol)
                  for (k in 1:length(tmpind)) {
                    if (missing(time.points)) {
                      lines(tmp[tmpind[k], ], col = colo[jj])
                    }
                    else lines(time.points, tmp[tmpind[k], ], 
                      col = colo[jj])
                  }
                }
            }
        }
        if (centre) {
            lines(cl[[1]][j, ], col = centre.col, lwd = centre.lwd)
        }
        if (single) 
            return()
    }
}

# Overview of all clusters
my_mfuzz.plot2(dgelist_log2cpm_mito_ordered, cl = cmeans_res, mfrow = c(2, 2),
               x11 = FALSE, xlab = "Samples",
               centre = TRUE, centre.col = "gray50", centre.lwd = 0.05)


## Check the genes present in each individual cluster
cmeans_res$cluster
cmeans_res$cluster[which(cmeans_res$cluster == 2)]

## Convert named vector to a table with genes per cluster
genes_per_cluster <-
  enframe(sort(cmeans_res$cluster), name = "Genes", value = "Cluster") %>%
  group_by(Cluster) %>%
  summarise(Genes = paste(Genes, collapse = ", "))

# Create a directory to save the results from the fuzzy clustering 
if (!dir.exists(here('analysis_output/fuzzy_clustering/'))) {
  dir.create(here('analysis_output/fuzzy_clustering/'), recursive = TRUE)}

write.table(genes_per_cluster, 
            file = here('analysis_output/fuzzy_clustering/genes_per_cluster.tsv'), 
            quote = TRUE, sep = "\t", row.names = FALSE, col.names = TRUE)

#
##
### Save the cluster plots
##
#


# Open the graphics device with options for saving individual files 
pdf(file = here("analysis_output/fuzzy_clustering/fuzzy_clusters_%03d.pdf"),
    onefile = FALSE)

# Apply plotting function
my_mfuzz.plot2(dgelist_log2cpm_mito_ordered, cl = cmeans_res, mfrow = c(1, 1),
                 x11 = FALSE, xlab = "Samples",
                 centre = TRUE, centre.col = "gray50", centre.lwd = 0.01)

# Close graphics device
dev.off()

```
